+++ sklearn/metrics/pairwise.py
@@ -25,9 +25,9 @@ from ..utils import (
     gen_batches,
     gen_even_slices,
     get_chunk_n_rows,
-    is_scalar_nan,
 )
 from ..utils._mask import _get_mask
+from ..utils._missing import is_scalar_nan
 from ..utils._param_validation import (
     Hidden,
     Interval,
